refactor: extract server boot sequence into a bootstrap service (#2839)
server/index.js mixed ~144 route mounts with ~600 lines of inline startup
orchestration. The boot work now lives in server/services/bootstrap.js as
three entry points — bootstrapServices() (migrations, collection-schema
verification, AI Toolkit + runner registration, spawner/autofixer/task
learning), runBootSequence() (background service inits + schedulers, then the
ordered instance/sync/media-queue/DB chain ending in httpServer.listen), and
registerShutdownHandlers() (the graceful-shutdown state machine).

Boot ordering, the DB fail-fast gate, and shutdown behavior are unchanged;
no code touches an AI provider at boot. index.js drops to route registration
plus three calls into the service.
