+++ .changelog/NEXT.md
@@ -81,3 +81,4 @@
 - **[issue-2890] Take the held-back `undici` 8 bump** — the failing `safeUrlFetch` pin test turned out not to be about `connect.lookup` at all (undici 8's connector still honors it, so the DNS-rebinding IP pin is expressed exactly as before). undici 8 dropped v7's implicit wrapping of the legacy `onConnect`/`onHeaders`/`onComplete` request-handler shape, so a bare undici-8 `Agent` handed to the *built-in* `fetch` of any Node that still bundles undici 7 (22/24/25, including the 24.x CI matrix) throws `InvalidArgumentError: invalid onRequestStart method` — and because `fetchGuarded` swallows network errors to null, every guarded fetch would have silently returned null rather than falling back to an unpinned connect. The new exported `buildPinnedDispatcher()` wraps the pinned `Agent` in undici 8's documented `Dispatcher1Wrapper`, which both the legacy and the v8 handler contracts accept, so one build works across Node 22→26. The end-to-end pin proof now runs through that production factory instead of a hand-rolled `new Agent(...)`, so it also guards the runtime-compat half.
 - **[issue-2881] Extract a shared cloud-CLI image-provider resolver** — `resolveCloudProviderConfig(settings, mode)` in the new `server/services/imageGen/cloudProviderConfig.js` owns the enable-gate + job-param bundle that seven dispatch sites (the image-gen route, the dispatcher's `generateImage`/`checkConnection`, first-pass gen, LoRA dataset gen, universe batch render, character reference sheets, pipeline visual stages) each duplicated once per cloud backend. `resolveMode` in `visualStageHelpers.js` becomes a candidate walk over an `isModeUsable` predicate instead of a ladder that grew two `if`s per backend, and the two providers' no-image error narrations now share `buildNoImageReason`. Behavior-preserving except that a queued grok render's response now reports `model: "grok-imagegen"` (it previously reported nothing) and a codex queue row records the effective model instead of an unset one — both already the value the provider renders with. The provider-module lifecycle harness (grok.js structurally copying codex.js) stays deferred until a third cloud CLI arrives, per the issue's acceptance criteria.
 - **[issue-2847] Confirm and harden the CyberCity intel tab bar** — the hand-rolled `role="tablist"` in `CityIntelPane` stays hand-rolled (its cyan/`font-pixel` HUD skin and severity-tinted count chips would need a one-off variant in the shared `ui/TabPills`, which no other call site would use) and now carries a comment saying so, so future audits skip it. The ARIA it duplicates is corrected in the process: the collapse toggle moves out of the tablist (it was an invalid non-tab child), each tab gains `id`/`aria-controls` paired to a real `role="tabpanel"`, roving `tabIndex` plus Left/Right/Home/End keyboard navigation, and `aria-expanded` on the collapse toggle; `aria-selected` now tracks the selected tab rather than dropping to false while the pane is collapsed. The three buttons are driven from the already-exported `CITY_INTEL_TABS` instead of triplicated markup.
+- **[issue-2839] Extract the boot sequence from `server/index.js` into a bootstrap service** — the ~600 lines of inline startup orchestration (data migrations, collection-schema verification, AI Toolkit construction + runner registration, ~40 fire-and-forget service inits/schedulers, the ordered instance/sync/media-queue/DB chain that ends in `httpServer.listen()`, and the graceful-shutdown state machine) move to `server/services/bootstrap.js` as `bootstrapServices()` / `runBootSequence()` / `registerShutdownHandlers()`. `index.js` drops from 1259 to ~460 lines and is now HTTP/Socket.IO construction, route registration, and three calls into the bootstrap service. Boot ordering, the DB fail-fast gate, and shutdown behavior are unchanged; no new LLM calls at boot.
